Added all the exploits, the exploits that didn't work have been deleted, and the ones that works have been renamed.
The chains that contains LazyMap.get can all be exploitable, the chains containing DefaultedMap.get can execute che calculator but doesn't follow the specific chain inside their CommonsCollection file.

The only chains that are not exploitable in any way are the ones starting with the ReferenceIdentityMap object, because that class uses another method to calculate the hash that does not contain hashes.